Gave 8 Ball Shield Generators more time. dPad heading. Parallax Threshold to NetworkTable. setPersistent works.

+++ src/main/kotlin/org/team2471/frc/lib/input/Controller.kt
@@ -67,7 +67,7 @@ open class Controller(val port: Int) {
     }
 
     val povDirection : Angle
-        get() = ds.getStickPOV(port, 0).degrees
+        get() = ds.getStickPOV(port, 0).toFloat().degrees
 
     var rumble: Double = 0.0
         set(value) {
